requirements.txt
selenium
undetected-chromedriver
python-dotenv
pymongo
pandas
openpyxl
pyvirtualdisplay
setuptools 